+++ src/app/projects/project/graph/components/node-components.type.ts
@@ -2,4 +2,5 @@ export enum NodeComponentsType {
     FileInput = 'FileInput',
     FileOutput = 'FileOutput',
     Script = 'Script',
+    TextInput = 'TextInput'
 }