+++ src/app/projects/project/graph/components/text-input/properties/text-input-properties.module.ts
@@ -1,0 +1,34 @@
+import { NgModule } from '@angular/core';
+import { ReactiveFormsModule } from '@angular/forms';
+import { BrowserModule } from '@angular/platform-browser';
+import { TranslateModule } from '@ngx-translate/core';
+import { InputTextModule } from 'primeng/inputtext';
+import { SelectButtonModule } from 'primeng/selectbutton';
+import { ProjectsServiceModule } from 'src/app/projects-service/projects-service.module';
+import { ProjectGraphServiceModule } from '../../../service/project-graph-service.module';
+import { TextInputPropertiesComponent } from './text-input-properties.component';
+
+@NgModule({
+    imports: [
+        // Angular
+        BrowserModule,
+        ReactiveFormsModule,
+        // ngx-translate,
+        TranslateModule,
+        SelectButtonModule,
+        // PrimeNG
+        InputTextModule,
+        // Custom
+        ProjectsServiceModule,
+        ProjectGraphServiceModule,
+    ],
+    declarations: [
+        TextInputPropertiesComponent
+    ],
+    exports: [
+        TextInputPropertiesComponent
+    ]
+})
+export class TextInputPropertiesModule {
+
+}