+++ src/app/projects/project/graph/components/text-input/text-input.data.ts
@@ -1,0 +1,7 @@
+import { SquishyNodeData } from '../squishy-node.data';
+import { TextInputType } from './text-input.type';
+
+export interface TextInputData extends SquishyNodeData {
+    content: string | undefined
+    inputType: TextInputType
+}