+++ src/app/projects/project/graph/components/text-input/text-input.type.ts
@@ -1,0 +1,9 @@
+export enum TextInputType {
+    PlainText = 'plaintext',
+    Json = 'json'
+}
+
+export const TextInputTypes: TextInputType[] = [
+    TextInputType.PlainText,
+    TextInputType.Json
+]